feat(tx-flow): move new ConfirmProposedTx flow components to replace the existing ones
