--- apps/web/src/components/tx-flow-2/flows/index.ts
@@ -1,3 +1,0 @@
-import dynamic from 'next/dynamic'
-
-export const ConfirmTxFlow = dynamic(() => import('./ConfirmTx'))
