+++ apps/web/src/components/tx-flow/flows/ConfirmTx/ConfirmProposedTx.tsx
@@ -2,16 +2,14 @@ import { type ReactElement, useContext, useEffect } from 'react'
 import { Typography } from '@mui/material'
 import { useChainId } from '@/hooks/useChainId'
 import { createExistingTx } from '@/services/tx/tx-sender'
-import { SafeTxContext } from '../../SafeTxProvider'
-import ReviewTransaction from '@/components/tx/ReviewTransaction'
+import ReviewTransaction from '@/components/tx/ReviewTransactionV2'
+import type { ReviewTransactionContentProps } from '@/components/tx/ReviewTransactionV2/ReviewTransactionContent'
+import { SafeTxContext } from '@/components/tx-flow/SafeTxProvider'
+import { TxFlowContext } from '@/components/tx-flow-2/TxFlowProvider'
 
 type ConfirmProposedTxProps = {
   txNonce: number | undefined
-  txId: string
-  onSubmit: () => void
-  isExecutable: boolean
-  onlyExecute: boolean
-}
+} & ReviewTransactionContentProps
 
 const SIGN_TEXT = 'Sign this transaction.'
 const EXECUTE_TEXT = 'Submit the form to execute this transaction.'
@@ -20,16 +18,19 @@ const SIGN_EXECUTE_TEXT = 'Sign or immediately execute this transaction.'
 const ConfirmProposedTx = ({ txNonce, ...props }: ConfirmProposedTxProps): ReactElement => {
   const chainId = useChainId()
   const { setSafeTx, setSafeTxError, setNonce } = useContext(SafeTxContext)
+  const { txId, onlyExecute, isExecutable } = useContext(TxFlowContext)
 
   useEffect(() => {
     txNonce !== undefined && setNonce(txNonce)
   }, [setNonce, txNonce])
 
   useEffect(() => {
-    createExistingTx(chainId, props.txId).then(setSafeTx).catch(setSafeTxError)
-  }, [props.txId, chainId, setSafeTx, setSafeTxError])
+    if (txId) {
+      createExistingTx(chainId, txId).then(setSafeTx).catch(setSafeTxError)
+    }
+  }, [txId, chainId, setSafeTx, setSafeTxError])
 
-  const text = !props.onlyExecute ? (props.isExecutable ? SIGN_EXECUTE_TEXT : SIGN_TEXT) : EXECUTE_TEXT
+  const text = onlyExecute ? (isExecutable ? SIGN_EXECUTE_TEXT : SIGN_TEXT) : EXECUTE_TEXT
 
   return (
     <ReviewTransaction {...props}>
