+++ apps/web/src/components/tx-flow/flows/ConfirmTx/index.tsx
@@ -1,22 +1,17 @@
 import { isSwapOrderTxInfo } from '@/utils/transaction-guards'
 import type { TransactionSummary } from '@safe-global/safe-gateway-typescript-sdk'
-import TxLayout from '@/components/tx-flow/common/TxLayout'
-import type { TxStep } from '../../common/TxLayout'
 import ConfirmProposedTx from './ConfirmProposedTx'
 import { useTransactionType } from '@/hooks/useTransactionType'
 import SwapIcon from '@/public/images/common/swap.svg'
 import { useMemo } from 'react'
-import useTxStepper from '../../useTxStepper'
-import { ConfirmTxDetails } from '@/components/tx/ConfirmTxDetails'
 import { isExecutable, isMultisigExecutionInfo, isSignableBy } from '@/utils/transaction-guards'
 import { useSigner } from '@/hooks/wallets/useWallet'
 import useSafeInfo from '@/hooks/useSafeInfo'
-import { TxFlowType } from '@/services/analytics'
+import { createDefaultTxFlow } from '@/components/tx-flow-2/createTxFlow'
 
 const ConfirmTxFlow = ({ txSummary }: { txSummary: TransactionSummary }) => {
   const { text } = useTransactionType(txSummary)
   const isSwapOrder = isSwapOrderTxInfo(txSummary.txInfo)
-  const { step, nextStep, prevStep } = useTxStepper(undefined, TxFlowType.CONFIRM_TX)
   const signer = useSigner()
   const { safe } = useSafeInfo()
 
@@ -25,36 +20,21 @@ const ConfirmTxFlow = ({ txSummary }: { txSummary: TransactionSummary }) => {
   const canExecute = isExecutable(txSummary, signer?.address || '', safe)
   const canSign = isSignableBy(txSummary, signer?.address || '')
 
-  const commonProps = useMemo(
-    () => ({ txId, isExecutable: canExecute, onlyExecute: !canSign, showMethodCall: true }),
-    [txId, canExecute, canSign],
-  )
-
-  const steps = useMemo<TxStep[]>(
-    () => [
-      {
-        txLayoutProps: { title: 'Confirm transaction' },
-        content: <ConfirmProposedTx key={0} txNonce={txNonce} onSubmit={() => nextStep(undefined)} {...commonProps} />,
-      },
-      {
-        txLayoutProps: { title: 'Confirm transaction details', fixedNonce: true },
-        content: <ConfirmTxDetails key={1} onSubmit={() => {}} {...commonProps} />,
-      },
-    ],
-    [nextStep, commonProps, txNonce],
+  const TxFlow = useMemo(
+    () => createDefaultTxFlow<undefined>((props) => <ConfirmProposedTx txNonce={txNonce} {...props} />),
+    [txNonce],
   )
 
   return (
-    <TxLayout
-      subtitle={<>{text}&nbsp;</>}
+    <TxFlow
       icon={isSwapOrder && SwapIcon}
-      step={step}
-      onBack={prevStep}
+      subtitle={<>{text}&nbsp;</>}
+      txId={txId}
+      isExecutable={canExecute}
+      onlyExecute={!canSign}
       txSummary={txSummary}
-      {...(steps?.[step]?.txLayoutProps || {})}
-    >
-      {steps.map(({ content }) => content)}
-    </TxLayout>
+      showMethodCall
+    />
   )
 }
